back end/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic<2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
httpx>=0.24.0
python-multipart>=0.0.5